# Improve iterator performance/usability