* Add `Arena::len()` and `Arena::is_empty()`
* Implements `ExactSizeIterator`, `DoubleEndedIterator`, `FusedIterator` for `Iter`, `IterMut`, `IntoIter`.

NOTE: This Pull Request depends on #5