@gifnksm 2018-10-12
@fitzgen I've added benchmark test of `Iterator::collect`.

about 20~40% perf wins!

## master

```
collect/collect-small/100
                        time:   [535.43 us 538.81 us 542.28 us]
                        thrpt:  [184.41 Kelem/s 185.59 Kelem/s 186.77 Kelem/s]
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe
collect/collect-small/200
                        time:   [1.0770 ms 1.0851 ms 1.0929 ms]
                        thrpt:  [182.99 Kelem/s 184.31 Kelem/s 185.70 Kelem/s]

collect/collect-big/100 time:   [560.80 us 563.52 us 566.40 us]
                        thrpt:  [176.55 Kelem/s 177.46 Kelem/s 178.32 Kelem/s]
collect/collect-big/200 time:   [1.1554 ms 1.1659 ms 1.1757 ms]
                        thrpt:  [170.11 Kelem/s 171.54 Kelem/s 173.11 Kelem/s]
```

## This branch

```
collect/collect-small/100
                        time:   [450.08 us 452.50 us 455.06 us]
                        thrpt:  [219.75 Kelem/s 220.99 Kelem/s 222.18 Kelem/s]
                        change: [-17.276% -16.143% -15.179%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 19 outliers among 100 measurements (19.00%)
  5 (5.00%) low severe
  6 (6.00%) low mild
  5 (5.00%) high mild
  3 (3.00%) high severe
collect/collect-small/200
                        time:   [869.34 us 874.32 us 880.32 us]
                        thrpt:  [227.19 Kelem/s 228.75 Kelem/s 230.06 Kelem/s]
                        change: [-18.866% -18.247% -17.602%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe

collect/collect-big/100 time:   [349.68 us 352.03 us 354.84 us]
                        thrpt:  [281.82 Kelem/s 284.07 Kelem/s 285.98 Kelem/s]
                        change: [-37.691% -36.793% -35.736%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 16 outliers among 100 measurements (16.00%)
  2 (2.00%) low mild
  5 (5.00%) high mild
  9 (9.00%) high severe
collect/collect-big/200 time:   [684.03 us 688.40 us 693.70 us]
                        thrpt:  [288.31 Kelem/s 290.53 Kelem/s 292.38 Kelem/s]
                        change: [-40.542% -40.030% -39.508%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe
```