@fitzgen 2018-10-12
Nice work! Will merge once the conflicts are fixed :)