@gifnksm 2018-10-13
Rebased on master. Ready for merge now!